refactor: DefaultType 구현
