+++ src/main/java/chess/model/Type.java
@@ -1,0 +1,5 @@
+package chess.model;
+
+public interface Type {
+
+}
