+++ src/main/java/chess/model/board/DefaultType.java
@@ -1,0 +1,7 @@
+package chess.model.board;
+
+import chess.model.Type;
+
+public enum DefaultType implements Type {
+    EMPTY
+}
