+++ src/main/java/chess/model/board/PieceSquare.java
@@ -1,7 +1,7 @@
 package chess.model.board;
 
+import chess.model.Type;
 import chess.model.piece.Piece;
-import chess.model.piece.Type;
 
 public class PieceSquare implements Square {
