+++ src/test/java/chess/model/board/PieceFixture.java
@@ -2,12 +2,12 @@ package chess.model.board;
 
 import chess.model.piece.Color;
 import chess.model.piece.Piece;
-import chess.model.piece.Type;
+import chess.model.piece.PieceType;
 
 public class PieceFixture {
-    public static final Piece WHITE_BISHOP = new Piece(Color.WHITE, Type.BISHOP);
-    public static final Piece BLACK_KNIGHT = new Piece(Color.BLACK, Type.KNIGHT);
-    public static final Piece WHITE_PAWN = new Piece(Color.WHITE, Type.PAWN);
+    public static final Piece WHITE_BISHOP = new Piece(Color.WHITE, PieceType.BISHOP);
+    public static final Piece BLACK_KNIGHT = new Piece(Color.BLACK, PieceType.KNIGHT);
+    public static final Piece WHITE_PAWN = new Piece(Color.WHITE, PieceType.PAWN);
 
 
     public static final Position A1 = new Position(File.A, Rank.FIRST);
